[ refactor ] 代辦項目元件檔名改為 CheckItems (複數)
